+++ components/ui/icons/activity.tsx
@@ -1,0 +1,79 @@
+/**
+ * Activity and Action Icons
+ */
+
+export interface IconProps {
+  className?: string;
+}
+
+export const ActivityIcon = ({ className = "w-5 h-5" }: IconProps) => (
+  <svg
+    className={className}
+    fill="none"
+    stroke="currentColor"
+    viewBox="0 0 24 24"
+  >
+    <path
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth="2"
+      d="M9 19v-6a2 2 0 00-2-2H5a2 2 0 00-2 2v6a2 2 0 002 2h2a2 2 0 002-2zm0 0V9a2 2 0 012-2h2a2 2 0 012 2v10m-6 0a2 2 0 002 2h2a2 2 0 002-2m0 0V5a2 2 0 012-2h2a2 2 0 012 2v14a2 2 0 01-2 2h-2a2 2 0 01-2-2z"
+    />
+  </svg>
+);
+
+export const RefreshCwIcon = ({ className = "w-4 h-4" }: IconProps) => (
+  <svg
+    className={className}
+    fill="none"
+    stroke="currentColor"
+    viewBox="0 0 24 24"
+  >
+    <path
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth="2"
+      d="M23 4v6h-6M1 20v-6h6"
+    />
+    <path
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth="2"
+      d="M3.51 9a9 9 0 0114.85-3.36L23 10M1 14l4.64 4.36A9 9 0 0020.49 15"
+    />
+  </svg>
+);
+
+export const RotateCcwIcon = ({ className = "w-5 h-5" }: IconProps) => (
+  <svg
+    className={className}
+    fill="none"
+    stroke="currentColor"
+    viewBox="0 0 24 24"
+  >
+    <path
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth="2"
+      d="M1 4v6h6M23 20v-6h-6M20.49 15a9 9 0 11-2.12-9.36L23 10"
+    />
+  </svg>
+);
+
+export const Loader2Icon = ({
+  className = "w-4 h-4 animate-spin",
+}: IconProps) => (
+  <svg
+    className={className}
+    fill="none"
+    stroke="currentColor"
+    viewBox="0 0 24 24"
+  >
+    <path
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth="2"
+      d="M12 2v4m0 12v4M2 12h4m12 0h4m-2.93-6.07l-2.83 2.83m-5.66 5.66l-2.83 2.83m2.83-11.32l2.83 2.83m-2.83 8.48l2.83 2.83"
+    />
+  </svg>
+);
